Fixed Rubocop Circle CI issue with configuration
Necessary to prevent Circle CI from failing the build due to using
`AllCops`. This might be a bug with Rubocop but can revert back to the
more explicit excludes to satisfy Circle CI. None of this is
recreatable locally, though. The Rubocop bug appears to be with this
exception:

    Running RuboCop...
    Unable to find gem panolint; is the gem installed?
    Gem::MissingSpecError
    /root/project/vendor/bundle/ruby/3.0.0/gems/rubocop-1.22.1/lib/
    rubocop/config_loader_resolver.rb:278:in `rescue in
    gem_config_path'
    /root/project/vendor/bundle/ruby/3.0.0/gems/rubocop-1.22.1/lib/
    rubocop/config_loader_resolver.rb:268:in `gem_config_path

This is a partial revert of this commit: 51bdf97c15c9.